feat: added auth

+++ features/Admin/Actions/deletePizzaAction.ts
@@ -2,9 +2,11 @@
 
 import { revalidatePath } from "next/cache";
 import { deletePizzaDal } from "../Dal/pizzaDal";
+import { requireAdmin } from "@/lib/requireAdmin";
 
 export async function deletePizzaAction(pizzaId: string) {
   try {
+    await requireAdmin();
     await deletePizzaDal(pizzaId);
 
     revalidatePath("/pizzas");
